move client config
