+++ src/config.ts
@@ -1,10 +1,3 @@
-export const clientConfig = {
-  apiKey: process.env.NEXT_PUBLIC_FIREBASE_API_KEY,
-  authDomain: process.env.NEXT_PUBLIC_FIREBASE_AUTH_DOMAIN,
-  projectId: process.env.NEXT_PUBLIC_FIREBASE_PROJECT_ID,
-  appId: process.env.NEXT_PUBLIC_FIREBASE_APP_ID,
-};
-
 export const serverConfig = {
   useSecureCookies: process.env.USE_SECURE_COOKIES === "true",
   firebaseApiKey: process.env.NEXT_PUBLIC_FIREBASE_API_KEY!,
